后台配置表

+++ database/migrations/2019_07_09_174952_create_courses_table.php
@@ -11,8 +11,8 @@ class CreateCoursesTable extends Migration
             $table->increments('id');
             $table->string('title')->index();
             $table->text('banner');
-            $table->integer('ori_price')->unsigned()->default(0);
-            $table->integer('now_price')->unsigned()->default(0);
+            $table->decimal('ori_price', 5, 2)->unsigned()->default(0);
+            $table->decimal('now_price', 5, 2)->unsigned()->default(0);
             $table->text('body');
             $table->integer('view_count')->default(0);
             $table->integer('buy_count')->index()->default(0);
